+++ package/components/StorkEntryTitle.svelte
@@ -1,16 +1,11 @@
-<script lang="ts">
-    import type {IStorkEntry, IStorkHighlightRange} from "../stores/search";
-    import Highlight from "./StorkHighlight.svelte";
-
-    export let element: HTMLParagraphElement | null = null;
-
-    let _class: string = "";
-    export let style: string | undefined = undefined;
-
-    export {_class as class};
-
-    export let entry: IStorkEntry;
-    export let title_highlight_ranges: IStorkHighlightRange[];
+<script lang="ts">;
+import Highlight from "./StorkHighlight.svelte";
+export let element = null;
+let _class = "";
+export let style = undefined;
+export { _class as class };
+export let entry;
+export let title_highlight_ranges;
 </script>
 
 <p bind:this={element} class="svst-entry-title {_class}" {style}>