+++ package/components/StorkExcerpt.svelte
@@ -1,16 +1,10 @@
-<script lang="ts">
-    import type {IStorkExcerpt} from "../stores/search";
-
-    import Highlight from "./StorkHighlight.svelte";
-
-    export let element: HTMLParagraphElement | null = null;
-
-    let _class: string = "";
-    export let style: string | undefined = undefined;
-
-    export {_class as class};
-
-    export let excerpt: IStorkExcerpt;
+<script lang="ts">;
+import Highlight from "./StorkHighlight.svelte";
+export let element = null;
+let _class = "";
+export let style = undefined;
+export { _class as class };
+export let excerpt;
 </script>
 
 <p bind:this={element} class="svst-excerpt {_class}" {style}>