+++ svelte.config.js
@@ -3,8 +3,4 @@ import sveltePreprocess from "svelte-preprocess";
 /** @type {import('@sveltejs/kit').Config} */
 export default {
     preprocess: [sveltePreprocess()],
-
-    kit: {
-        target: "#svelte",
-    },
 };